pyproject.toml
[tool.poetry]
name = "dive-server"
version = "0.1.0"
description = "Application server for DIVE"
authors = ["Bibek Pandey <bewakepandey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.6"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
django-cors-headers = "^3.13.0"
django-environ = "^0.8.1"
graphene-django = ">=2,<3"
graphene-graphiql-explorer = "*"
graphene-django-extras = { git = "https://github.com/the-deep/graphene-django-extras.git", branch = "v0.4.9-patch.2" }
graphene-file-upload = "^1.3.0"
django-enumfield = "^3.0"
factory-boy = "*"

[tool.poetry.dev-dependencies]
typing-extensions = "^4.3.0"

[tool.poetry.group.dev.dependencies]
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}
mypy = "^0.930"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
